feat: integrate Cobra CLI framework and restructure main pipeline
