+++ go.mod
@@ -5,3 +5,9 @@ go 1.24.2
 require gopkg.in/yaml.v3 v3.0.1
 
 require github.com/beevik/etree v1.5.1
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/cobra v1.9.1 
+	github.com/spf13/pflag v1.0.6 // indirect
+)
